Source/backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dataclasses-json==0.5.14
DateTime==5.2
dnspython==2.4.1
exceptiongroup==1.1.2
fastapi==0.101.0
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
langchain==0.0.260
langsmith==0.0.20
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pdfkit==1.0.0
Pillow==10.0.0
pydantic==1.10.12
pydantic_core==2.4.0
pymongo==4.4.1
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
reportlab==4.0.4
repoze.lru==0.7
requests==2.31.0
Routes==2.5.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tenacity==8.2.2
tqdm==4.66.1
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
wkhtmltopdf==0.2
yarl==1.9.2
zope.interface==6.0
pdf2image~=1.16.3
